% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{beta_dist}
\alias{beta_dist}
\title{Taxonimic betadiversity among several communities}
\usage{
beta_dist(W)
}
\arguments{
\item{W}{matrix containing the abundances of the species per site (or presence-absence,
i.e. 0 or 1). Rows are sites and species are columns. NA not tolerated. \code{}}
}
\value{
Btot Total Beta diversity dist object

B_3 Beta diversity due to replacement dist object

Brich Beta diversity due to richness diferences dist object

quality the ouptput will print the quality of the dendogram representation.
clustering  performance is assessed by the correlation with the cophenetic distance
}
\description{
Calculate taxonomic trait beta diversity among all pairwise communities using Carvalho et al. 2012 decomposition
and creates a dist object.
}
\examples{
ex1 <- beta_dist(W = dummy$abun)
ex1
}

