% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beta_part.R
\name{beta_part}
\alias{beta_part}
\title{Taxonimic betadiversity}
\usage{
beta_part(c1, c2)
}
\arguments{
\item{c1}{vector containing the species names in the first community. NA not tolerated. \code{}}

\item{c2}{vector containing the species names in the first community. NA not tolerated. \code{}}
}
\value{
Btot Total Beta diversity

B_3 Beta diversity due to replacement

Brich Beta diversity due to richness diferences

quality the ouptput will print the quality of the dendogram representation.
clustering  performance is assessed by the correlation with the cophenetic distance
}
\description{
Calculate taxonomic trait beta diversity for two communities using Carvalho et al. 2012 decomposition.
}
\examples{
ex1 <- beta_part(c1 = c("sp3", "sp2", "sp1", "sp4", "sp5"), c2 = c("sp6", "sp7", "sp8", "sp4", "sp5"))
ex1
}

