% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/betaFD.R
\name{betaFD}
\alias{betaFD}
\title{Functional betadiversity}
\usage{
betaFD(c1, c2, S, Tree = NULL)
}
\arguments{
\item{c1}{vector containing the species names in the first community. NA not tolerated. In addition,
the species labels in c1 and S must be identical. \code{}}

\item{c2}{vector containing the species names in the first community. NA not tolerated. In addition,
the species labels in c2 and S must be identical.\code{}}

\item{S}{matrix or data frame of functional traits. Traits can be numeric, ordered,
or factor. NAs are tolerated. All species in c1 and c2 should be present with no extra species.
Species should be ordered alphabetically.\code{}}
}
\value{
Btot Total Betadiversity

B_3 Beta diversity due to replacement

Brich Beta diversity due to richness diferences

quality the ouptput will print the quality of the dendogram representation.
clustering  performance is assessed by the correlation with the cophenetic distance
}
\description{
Calculate functional trait beta diversity for two communities using Petchey and Gaston 2002 index
and Carvalho et al. 2012 decomposition.
}
\examples{
ex1 <- betaFD(c1 = c("sp3", "sp2", "sp1", "sp4", "sp5"), c2 = c("sp6", "sp7", "sp8", "sp4", "sp5"), S = dummy$trait)
ex1
}

