% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/betaFD_dist.R
\name{betaFD_dist}
\alias{betaFD_dist}
\title{Functional betadiversity among several communities}
\usage{
betaFD_dist(W, S)
}
\arguments{
\item{W}{matrix containing the abundances of the species in S (or presence-absence,
i.e. 0 or 1). Rows are sites and species are columns. NA not tolerated.
In addition, the species labels in W and S must be identical.\code{}}

\item{S}{matrix or data frame of functional traits. Traits can be numeric, ordered,
or factor. NAs are tolerated. All species in W should be present.
Species should be ordered alphabetically.\code{}}
}
\value{
Btot Total Betadiversity dist object

B_3 Beta diversity due to replacement dist object

Brich Beta diversity due to richness diferences dist object

quality the ouptput will print the quality of the dendogram representation.
clustering  performance is assessed by the correlation with the cophenetic distance
}
\description{
Calculate functional trait beta diversity for several communities using Petchey and Gaston 2002 index
and Carvalho et al. 2012 decomposition and creates a dist object.
}
\examples{
ex1 <- betaFD_dist(W = dummy$abun, S = dummy$trait)
ex1
}

