% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Fredundancy.R
\name{FRedundancy}
\alias{FRedundancy}
\title{Calculate Functional Reduncancy}
\usage{
FRedundancy(S, A, it = 10)
}
\arguments{
\item{S}{matrix or data frame of functional traits. Traits can be numeric, ordered,
or factor. NAs are tolerated.\code{}}

\item{A}{matrix containing the abundances of the species in x (or presence-absence,
i.e. 0 or 1). Rows are sites and species are columns. NA not tolerated. The number of
species (columns) in a must match the number of species (rows) in x. In addition,
the species labels in a and x must be identical and in the same order.\code{}}

\item{it}{number of iterations}
}
\value{
site vector of sites

rob_Fpg vector of robustness (indicates reduncancy) using petchey and Gaston index

rob_FRich vector of robustness using distance-Based indexes (index) Laliberte
}
\description{
work in progress to calculate FR ...
}

