% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functcomp_bm.R
\name{functcomp_bm}
\alias{functcomp_bm}
\title{Functional Composition from FD package with the option to weith by biomass}
\usage{
functcomp_bm(x, a, CWM.type = c("dom", "all"), bin.num = NULL,
  Weigthedby = c("abundance", "biomasCarabids", "biomasBees", "biomassValue"),
  biomassValue = NA)
}
\arguments{
\item{Weigthedby}{character string indicating if should be weighted by `abundance`
or `biomassValue`. If biomassValue is in length units for Carabids or bees,
use options `biomasCarabids` or `biomasBees` to automatically convert it to mass.\code{}}

\item{biomassValue}{numerical vector with body weigh (or length) values for each species
in the same order as species are provided. It can also be a matrix or data
frame with one mass value for each community and species (both communities and species
arranged like in a). \code{}}

\item{same}{as functcomp\code{}}
}
\value{
a data frame containing the CWM values (weighted by abundance or by biomass) of each trait for each community.\code{}
}
\description{
functcomp returns the functional composition of a set of communities, as measured by the community-level weighted means of trait values (CWM; e.g. Lavorel et al. 2008).
functcomp_w also allows to weight CWM by biomass
}
\examples{
ex1 <- functcomp_bm(dummy$trait, dummy$abun, Weigthedby = "biomassValue",
biomassValue = c(1.2, 2.3, 0.6, 1.0, 3.4, 0.2, 1.6, 2.2))
ex1
}

