% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/null.FD.R
\name{null.FD}
\alias{null.FD}
\title{Calculate Null Model of FD values obtained under random species richness for a given community}
\usage{
null.FD(S, A, it, w = NA)
}
\arguments{
\item{S}{matrix or data frame of functional traits. Traits can be numeric, ordered,
or factor. NAs are tolerated.\code{}}

\item{A}{matrix containing the abundances of the species in x (or presence-absence,
i.e. 0 or 1). Rows are sites and species are columns. NA not tolerated. The number of
species (columns) in a must match the number of species (rows) in x. In addition,
the species labels in a and x must be identical and in the same order.\code{}}

\item{it}{number of iterations}

\item{w}{vector listing the weights for the traits in x. Can be missing,
in which case all traits have equal weights.}
}
\value{
comm vector listing the evaluated communities

Rich vector with the observed species richness

FD  vector with the observed FD index

null_meanFD vector with the expected mean FD value for that richness
level under random conditions

null_sdFD vector with the S.E. of the FD value for that richness
level under random conditions

pFD  p-value comparing FD and its expected null

Frich vector with the observed FD index

null_meanFrich vector with the expected mean FD value for that richness
level under random conditions

null_sdFrich vector with the S.E. of the FD value for that richness
level under random conditions

pFrich p-value comparing FD and its expected null
}
\description{
work in progress to calculate FR ...
}

