\name{Xtree}
\alias{Xtree}
\title{Evaluate species branch matrix (sensu Petchey&Gaston) from a dendrogram}
\usage{
  Xtree(h)
}
\arguments{
  \item{h}{a dendrogram generated by hclust()\code{}}
}
\value{
  h2.prime contains the length of all the tiny branches

  H1 each row represents one species, each column
  represents one branch 1 indicates that a branch is part
  of the pathway from species to top of the dendrogram 0
  otherwise
}
\description{
  17/1/03. Written by Jens Schumacher. Please acknowledge
  as appropriate. tested for results of hclust and agnes
  hclust - hierarchical clustering agnes - agglomerative
  clustering used components: merge - history of cluster
  merging height - actual heights at merging order -
  permutation to achieve nice output (needed only for
  agnes)
}

