\docType{package}
\name{fundiv-package}
\alias{fundiv}
\alias{fundiv-package}
\title{fundiv: analyzing functional trait diversity}
\description{
  tools to calculate several functional trait diversity
  indexes
}
\details{
  \tabular{ll}{ Package: \tab fundiv\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2013-10-22\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr}
}
\examples{
#see functions.
}
\author{
  Ignasi Bartomeus \email{nacho.bartomeus@gmail.com}
}
\keyword{package}

